Fix: Use year-to-date Fantrax regular stats
